shared/pyproject.toml
[project]
name = "tba-shared"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-logging",
    "google-cloud-ndb",
    "google-cloud-secret-manager",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared"]
